backend/requirements.txt
nltk==3.6.2
pandas==1.3.2
fastai==2.5.2
numpy==1.21.2
requests==2.26.0
Flask==2.0.1
Pillow==8.3.2
